+++ registry/bentos/home-screen/main.go
@@ -22,7 +22,7 @@ import (
 	"github.com/cloudboy-jh/bentotui/theme"
 )
 
-const version = "v0.5.4"
+const version = "v0.6.0"
 
 const wordmark = "" +
 	"██████╗ ███████╗███╗   ██╗████████╗ ██████╗ \n" +